backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.69
xgboost==2.0.3
torch==2.3.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
ultralytics==8.2.0
